Board(#2), Tests(#5): Added flood fill accessibility checking method and tests.
The accessibility checking method allows the discovery of all points reachable
from a given point in the map.
Added utility method to read in a map as a string instead of as a board.
Board toString method modified to use an additional method which allows the
passing of a list to mark points in it. toString calls this method with a null list.
Tests for the flood fill check a map for accessibility by comparing the string
representations.