Tests(#5), Board(#2): Possible push actions method, accessibility modified.
- New method which returns the possible push actions which can be applied to each box.
- The accessibility function now keeps track of the top left most point in the accessible area.
The top left most point prioritises the y value of the point, and then the x. So, if the accessible
area contains the point (0,1), if there is a point where the y coordinate is 0, then that point will
be the top left most point no matter what the x value of the point (this should probably be changed).
- Tests updated for the accessibility function and added for the possible push actions.
- Board constructor automatically finds the accessible area and stores the top left point as a member.