Board(#2), Tests(#5): Board equality and tests.
- Implemented a new equality test for boards which ignores the exact position
of the player. For two boards to be equal, they must have boxes in the
same places, and the top left most position of the accessible area of the
player must be the same for both boards. This version of equality is not
currently portable, and as such will not be used in search methods, as the
equals method of the generic object is used. Probably need to stop using generics
and implement the searches using the classes that are actually used so that board
expansion can be done correctly.
- Testing for the new equality, updated testBuildFullPath test to use the new method.
- More maps for testing.