Search(#4): Modified SearchNode implementation.
- SearchNode constructor modified. Now takes a boolean parameter which determines
whether this node stores boards which are exploring the board space or the player
motion space.
- The new boolean parameter is used in the equals and expand method to determine
which equals/expand method of the Board is supposed to be called.
- Added a test case for the board space search. Needs an implementation of the
board space expansion to work.
- Modified the search methods so that they pass on the boolean parameter they are
called with to the search nodes that they construct.
- Goal boards are now put into SearchNode objects so that goal checking can be done
in the expected way through the SearchNode equals method which takes into account
which space is being searched.